make SIMPoroElasticity a private template
